+++ src/gui/NavBarML.ts
@@ -12,7 +12,7 @@ let navHTML = `
 
                 <button id="downWorld" title="download scene"><span class="material-icons-outlined" >cloud_download</span></button>
 
-                <button id="navWorldAssets" title="list items in scene"><span class="material-icons-outlined">account_tree</span></button>
+                <button id="navWorldAssets" title="list items in world"><span class="material-icons-outlined">account_tree</span></button>
 
                 <button id="navAllAssets" title="all files"><span class="material-icons-outlined" >folder</span></button>